Apply the review of PR 1: one rule per argument kind, and one home each
Sequence arguments (site ids, names, member indices) are sequences, always:
one bare id or one string is a TypeError naming the fix, a set is refused,
order is kept, and dict.keys() and NumPy, JAX, xarray and pandas arrays are
accepted, a 0-d one for a single id. allow_one_id is gone. Integers refuse
floats, integral or not, and a site the data lacks is a KeyError.
validation raises only through check_* helpers whose messages name a fix,
and as_batched_flat returns the 2-D batch alone (float64 for NumPy and JAX,
tracer lists under jit, None refused); ForwardModel checks its own rows and
finiteness. as_names now serves every names argument.

FrozenMapping moves to conventions as a dict subclass whose mutators raise,
and replaces every MappingProxyType; the attribute dicts are handed to
xarray as they are. SITE_ATTRIBUTES no longer names the 8000-site pool.
N_SITES lives once, in sites, for the ingest scripts and expected_rows.
conventions constants are no longer re-exported (constraints.TIME_BOUNDS_*,
CF_CONVENTIONS, sites.DATA_ROOT_ENV_VAR, the initial_conditions package's
SITE, NAME_PATTERN, CF_CONVENTIONS); conventions gains TIME_BOUNDS.

sites splits its site-table check into one-invariant checks with two
groups, adds site_coordinates (site, lon, lat) and the pool checks the raw
data sources share, and site_lookup type-checks its table. The parameter
vector uses them, reads site_labels.LABEL_COLUMN, reserves sample, point,
x, y, lat and lon, hands out writable lon/lat copies, and states which
vectors pickle. load_drivers keeps the requested site order. projection and
the 16-class split write their file pairs through write_checked_together.
Tracked raw inputs are found from the repository in tests and in the
Natural Earth scripts. Literals give way to the conventions constants and
pySIPNET's BOUNDS_DIMENSION, duplicated helpers go to validation and sites,
and import order and long lines are fixed on the lines this PR touched.

+++ src/sipnet_calibration/plotting/primitives.py
@@ -162,7 +162,7 @@ def spaghetti(
 
     label = style.pop("label", None)
     drawn = []
-    for position, index in enumerate(thinned_indices(len(samples), int(n_max))):
+    for position, index in enumerate(thinned_indices(len(samples), n_max)):
         keep_label = position == 0 and label is not None
         drawn.append(
             line(
@@ -440,8 +440,8 @@ def site_cells(
     if not (x_max > x_min and y_max > y_min):
         raise ValueError(f"bounds must have positive width and height, got {bounds}")
 
-    size = (x_max - x_min) / int(pixels)
-    n_x, n_y = int(pixels), max(1, int(np.ceil((y_max - y_min) / size)))
+    size = (x_max - x_min) / pixels
+    n_x, n_y = pixels, max(1, int(np.ceil((y_max - y_min) / size)))
     centers_x = x_min + size * (np.arange(n_x) + 0.5)
     centers_y = y_min + size * (np.arange(n_y) + 0.5)
     grid = np.stack(np.meshgrid(centers_x, centers_y), axis=-1).reshape(-1, 2)
